Construção de agendamento da janela de votação que vem por default 1min

+++ msvotacao/src/main/java/br/com/project/msvotacao/service/VotacaoService.java
@@ -5,6 +5,6 @@ import br.com.project.msvotacao.exception.ErroComunicacaoMicroservicesException;
 
 public interface VotacaoService {
 
-    VotacaoPostResponse abrirVotacao(int tempoParavotacao) throws ErroComunicacaoMicroservicesException;
+    VotacaoPostResponse abrirVotacao(long tempoParavotacao) throws ErroComunicacaoMicroservicesException;
 
 }
